backend/requirements.txt
fastapi
uvicorn
sse-starlette
python-multipart
scikit-learn
pdfplumber
PyMuPDF
PyPDF2
python-pptx
sentence-transformers
faiss-cpu

numpy
nltk
reportlab
bcrypt